+++ src/helpers/cli.ts
@@ -58,6 +58,11 @@ const args: ICLIArgument[] = [
     description: 'Target file(s) to upload',
   },
   {
+    name: 'preventSymbolicLinks',
+    type: 'boolean',
+    description: 'Specifies whether to prevent following of symbolic links. Defaults to false (symbolic links are followed).'
+  },
+  {
     name: 'fullReport',
     type: 'string',
     description: 'Specify the path to a previously uploaded Codecov report'