+++ src/helpers/web.ts
@@ -1,9 +1,6 @@
 import dns from 'node:dns'
 import { snakeCase } from 'snake-case'
-import {
-  request,
-  setGlobalDispatcher,
-} from 'undici'
+import { request, setGlobalDispatcher, errors, Dispatcher } from 'undici'
 
 import { version } from '../../package.json'
 import {
@@ -15,8 +12,14 @@ import {
   UploaderEnvs,
   UploaderInputs,
 } from '../types'
-import { info } from './logger'
+import { UploadLogger, info, logError } from './logger'
 import { addProxyIfNeeded } from './proxy'
+import { sleep } from './util'
+
+
+const maxRetries = 4
+const baseBackoffDelayMs = 1000 // Adjust this value based on your needs.
+export const userAgent: string = `codecov-uploader/${version}`
 
 /**
  *
@@ -33,7 +36,7 @@ export function populateBuildParams(
   serviceParams.name = args.name || envs.CODECOV_NAME || ''
   serviceParams.tag = args.tag || ''
 
-  if (typeof args.flags === "string") {
+  if (typeof args.flags === 'string') {
     serviceParams.flags = args.flags
   } else {
     serviceParams.flags = args.flags.join(',')
@@ -51,12 +54,38 @@ export function getPackage(source: string): string {
   }
 }
 
+async function requestWithRetry(
+  url: string,
+  options: Dispatcher.RequestOptions,
+  retryCount = 0,
+): Promise<Dispatcher.ResponseData> {
+  try {
+    const response = await request(url, options)
+    return response
+  } catch (error: unknown) {
+    if (
+      (
+        (error instanceof errors.UndiciError && error.code == 'ECONNRESET') ||
+        (error instanceof errors.UndiciError && error.code == 'ETIMEDOUT') ||
+        error instanceof errors.ConnectTimeoutError ||
+        error instanceof errors.SocketError
+      ) && retryCount < maxRetries
+    ) {
+      const backoffDelay = baseBackoffDelayMs * 2 ** retryCount
+      await sleep(backoffDelay)
+      UploadLogger.verbose('Request to Codecov failed. Retrying...')
+      logError(`Request error: ${error.message}`)
+      return requestWithRetry(url, options, retryCount + 1)
+    }
+    throw error
+  }
+}
 
 export async function uploadToCodecovPUT(
   putAndResultUrlPair: PostResults,
   uploadFile: string | Buffer,
   envs: UploaderEnvs,
-  args: UploaderArgs
+  args: UploaderArgs,
 ): Promise<PutResults> {
   info('Uploading...')
 
@@ -69,8 +98,11 @@ export async function uploadToCodecovPUT(
   if (requestHeaders.agent) {
     setGlobalDispatcher(requestHeaders.agent)
   }
-  dns.setDefaultResultOrder('ipv4first');
-  const response = await request(requestHeaders.url.origin, requestHeaders.options)
+  dns.setDefaultResultOrder('ipv4first')
+  const response = await requestWithRetry(
+    requestHeaders.url.origin,
+    requestHeaders.options,
+  )
 
   if (response.statusCode !== 200) {
     const data = await response.body.text()
@@ -79,7 +111,7 @@ export async function uploadToCodecovPUT(
     )
   }
 
-  return { status: 'success', resultURL: putAndResultUrlPair.resultURL }
+  return { status: 'processing', resultURL: putAndResultUrlPair.resultURL }
 }
 
 export async function uploadToCodecovPOST(
@@ -101,8 +133,12 @@ export async function uploadToCodecovPOST(
   if (requestHeaders.agent) {
     setGlobalDispatcher(requestHeaders.agent)
   }
-  dns.setDefaultResultOrder('ipv4first');
-  const response = await request(requestHeaders.url.origin, requestHeaders.options)
+  dns.setDefaultResultOrder('ipv4first')
+
+  const response = await requestWithRetry(
+    requestHeaders.url.origin,
+    requestHeaders.options,
+  )
 
   if (response.statusCode !== 200) {
     const data = await response.body.text()
@@ -113,7 +149,6 @@ export async function uploadToCodecovPOST(
 
   return await response.body.text()
 }
-
 /**
  *
  * @param {Object} queryParams
@@ -125,8 +160,6 @@ export function generateQuery(queryParams: Partial<IServiceParams>): string {
   ).toString()
 }
 
-
-
 export function parsePOSTResults(putAndResultUrlPair: string): PostResults {
   info(putAndResultUrlPair)
 
@@ -136,7 +169,9 @@ export function parsePOSTResults(putAndResultUrlPair: string): PostResults {
   const matches = putAndResultUrlPair.match(re)
 
   if (matches === null) {
-    throw new Error(`Parsing results from POST failed: (${putAndResultUrlPair})`)
+    throw new Error(
+      `Parsing results from POST failed: (${putAndResultUrlPair})`,
+    )
   }
 
   if (matches?.length !== 2) {
@@ -147,12 +182,12 @@ export function parsePOSTResults(putAndResultUrlPair: string): PostResults {
 
   if (matches[0] === undefined || matches[1] === undefined) {
     throw new Error(
-      `Invalid URLs received when parsing results from POST: ${matches[0]},${matches[1]}`
+      `Invalid URLs received when parsing results from POST: ${matches[0]},${matches[1]}`,
     )
   }
   const resultURL = new URL(matches[0].trimEnd())
   const putURL = new URL(matches[1])
-   // This match may have trailing 0x0A and 0x0D that must be trimmed
+  // This match may have trailing 0x0A and 0x0D that must be trimmed
 
   return { putURL, resultURL }
 }
@@ -170,13 +205,15 @@ export function generateRequestHeadersPOST(
   envs: UploaderEnvs,
   args: UploaderArgs,
 ): IRequestHeaders {
-  const url = new URL(`upload/v4?package=${getPackage(
-    source,
-  )}&token=${token}&${query}`, postURL)
+  const url = new URL(
+    `upload/v4?package=${getPackage(source)}&token=${token}&${query}`,
+    postURL,
+  )
 
-  const headers = {
+  const headers: Record<string, string> = {
     'X-Upload-Token': token,
     'X-Reduced-Redundancy': 'false',
+    'User-Agent': userAgent,
   }
 
   return {
@@ -191,17 +228,16 @@ export function generateRequestHeadersPOST(
   }
 }
 
-
-
 export function generateRequestHeadersPUT(
   uploadURL: URL,
   uploadFile: string | Buffer,
   envs: UploaderEnvs,
   args: UploaderArgs,
 ): IRequestHeaders {
-  const headers = {
+  const headers: Record<string, string> = {
     'Content-Type': 'text/plain',
     'Content-Encoding': 'gzip',
+    'User-Agent': userAgent,
   }
 
   return {