+++ test/helpers/files.test.ts
@@ -27,10 +27,10 @@ describe('File Helpers', () => {
   it('can fetch the git root', () => {
     const cwd = td.replace(process, 'cwd')
     const spawnSync = td.replace(childProcess, 'spawnSync')
-    td.when(cwd()).thenReturn({ stdout: 'fish' })
+    td.when(cwd()).thenReturn('fish')
     td.when(
       spawnSync('git', ['rev-parse', '--show-toplevel'], { maxBuffer: SPAWNPROCESSBUFFERSIZE }),
-    ).thenReturn({ stdout: 'gitRoot' })
+    ).thenReturn({ stdout: Buffer.from('gitRoot') })
 
     expect(fileHelpers.fetchGitRoot()).toBe('gitRoot')
   })
@@ -38,8 +38,8 @@ describe('File Helpers', () => {
   it('returns cwd when it cannot fetch the git root', () => {
     const cwd = td.replace(process, 'cwd')
     td.replace(childProcess, 'spawnSync')
-    td.when(cwd()).thenReturn({ stdout: 'fish' })
-    expect(fileHelpers.fetchGitRoot()).toEqual({stdout: 'fish'})
+    td.when(cwd()).thenReturn('fish')
+    expect(fileHelpers.fetchGitRoot()).toEqual('fish')
   })
 
   it('can get a file listing', async () => {
@@ -52,7 +52,7 @@ describe('File Helpers', () => {
       }
     })
     expect(
-      await fileHelpers.getFileListing('.', { flags: '', verbose: 'true', slug: '',upstream: '' }),
+      await fileHelpers.getFileListing('.', { flags: '', verbose: 'true', slug: '', upstream: '' }),
     ).toBe(files.join('\n'))
   })
 
@@ -277,6 +277,11 @@ describe('File Helpers', () => {
       it('should return path when project root is . and filepath starts /', () => {
         expect(fileHelpers.getFilePath('.', '/usr/coverage.xml')).toEqual(
           '/usr/coverage.xml',
+        )
+      })
+      it('should return path when file path starts with Windows letter drive', () => {
+        expect(fileHelpers.getFilePath('.', 'C:\Users\circleci\project\coverage\cobertura-coverage.xml')).toEqual(
+          'C:\Users\circleci\project\coverage\cobertura-coverage.xml',
         )
       })
     })